requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-js-asset==2.1.0
django-tabbed-admin==1.0.4
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
magic-filter==1.0.11
multidict==6.0.4
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
yarl==1.9.2
